+++ src/sprelay/core/k8090.cpp
@@ -20,18 +20,17 @@
 **                                                                        **
 ****************************************************************************/
 
-#include "k8090.h"
-
-#include <QSerialPort>
-#include <QSerialPortInfo>
-
-#include <QStringBuilder>
-
-#include <QDebug>
-
 /*!
     \file k8090.h
 */
+
+#include "k8090.h"
+
+#include <QDebug>
+#include <QSerialPort>
+#include <QSerialPortInfo>
+#include <QStringBuilder>
+
 
 /*!
     \namespace K8090Traits
@@ -40,29 +39,29 @@
 using namespace K8090Traits;  // NOLINT(build/namespaces)
 
 // static constants
-const quint16 K8090::productID = 32912;
-const quint16 K8090::vendorID = 4303;
+const quint16 K8090::kProductID = 32912;
+const quint16 K8090::kVendorID = 4303;
 
-const QString K8090::stxByte = "04";
-const QString K8090::etxByte = "0F";
-const QString K8090::switchRelayOnCmd = "11";
-const QString K8090::switchRelayOffCmd = "12";
-const QString K8090::toggleRelayCmd = "14";
-const QString K8090::queryRelayStatusCmd = "18";
-const QString K8090::setButtonModeCmd = "21";
-const QString K8090::queryButtonModeCmd = "22";
-const QString K8090::startRelayTimerCmd = "41";
-const QString K8090::setRelayTimerDelayCmd = "42";
-const QString K8090::queryTimerDelayCmd = "44";
-const QString K8090::buttonStatusCmd = "50";
-const QString K8090::relayStatusCmd = "51";
-const QString K8090::resetFactoryDefaultsCmd = "66";
-const QString K8090::jumperStatusCmd = "70";
-const QString K8090::firmwareVersionCmd = "71";
+const QString K8090::kStxByte_ = "04";
+const QString K8090::kEtxByte_ = "0F";
+const QString K8090::kSwitchRelayOnCmd_ = "11";
+const QString K8090::kSwitchRelayOffCmd_ = "12";
+const QString K8090::kToggleRelayCmd_ = "14";
+const QString K8090::kQueryRelayStatusCmd_ = "18";
+const QString K8090::kSetButtonModeCmd_ = "21";
+const QString K8090::kQueryButtonModeCmd_ = "22";
+const QString K8090::kStartRelayTimerCmd_ = "41";
+const QString K8090::kSetRelayTimerDelayCmd_ = "42";
+const QString K8090::kQueryTimerDelayCmd_ = "44";
+const QString K8090::kButtonStatusCmd_ = "50";
+const QString K8090::kRelayStatusCmd_ = "51";
+const QString K8090::kResetFactoryDefaultsCmd_ = "66";
+const QString K8090::kJumperStatusCmd_ = "70";
+const QString K8090::kFirmwareVersionCmd_ = "71";
 
 /*!
     \class K8090
-    \brief The class that provides the interface for %Velleman K8090 relay card
+    \brief The class that provides the interface for Velleman %K8090 relay card
     controlling through serial port.
 
     \remark reentrant, thread-safe
@@ -70,7 +69,7 @@ const QString K8090::firmwareVersionCmd = "71";
 
 // initialization of static member variables
 /*!
-    \brief Array of 4 byte (3 leading bytes and one command byte) commands used
+    \brief Array of 2 byte (1 leading byte and one command byte) commands used
     to control the relay.
 
     It is filled by fillCommandsArrays() static method, the first command is
@@ -81,42 +80,46 @@ const QString K8090::firmwareVersionCmd = "71";
     Example, which shows how to build the whole command read status:
 
     \code
-    int n = 6; // number of command bytes
+    int n = 7;  // Number of command bytes.
     unsigned char cmd[n]; // array of command bytes
 
     // copying the first two bytes of command to command byte array
-    for (int ii = 0; ii < 4; ++ii)
-        cmd[ii] = bCommands[static_cast<int>(Command::ReadStatus)][ii];
-    cmd[4] = 0; // 5th byte specifies the number of data bytes. For read
+    for (int ii = 0; ii < 2; ++ii)
+        cmd[ii] = b_commands_[static_cast<int>(Command::RELAY_STATUS)][ii];
+    cmd[2] = static_cast<unsigned char>(RelayID::ONE);  // 3rd byte specifies affected relays
     // commands, there is no one.
-    cmd[5] = K8090::checkSum(cmd, 5); // the last byte contains check sum.
+    cmd[5] = K8090::checkSum(cmd, 5); // sixth byte contains check sum.
+    cmd[6] = b_etx_byte_;
     \endcode
 */
-unsigned char K8090::bCommands[static_cast<int>(Command::None)][2];
-
-unsigned char K8090::bEtxByte;
+unsigned char K8090::b_commands_[static_cast<int>(Command::NONE)][2];
 
 /*!
-    \brief Array of QString representation of command bytes used to control the
-    bath.
+    \brief End delimiting byte.
+*/
+unsigned char K8090::b_etx_byte_;
 
-    To assemble the full command, it is necessary to prepend the commandBase
-    and append byte, which contains number of data bytes, folowed bytes
+/*!
+    \brief Array of QString representation of command bytes used to control
+    the relay card.
+
+    To assemble the full command, it is necessary to prepend the kStxByte
+    and append byte, which specifies the relays to be used, folowed by bytes
     containing data and ended with checksum (See CheckSum(const unsigned char,
-    int)). It is filled by fillCommandsArrays() static method, the first
-    command is the command with the most important priority, the last is the
-    least important. They should be accessed using the K8090Traits::Command
-    enum values. Example, which shows how to build the whole command Read
-    status:
+    int)) and kEtxByte. It is filled by fillCommandsArrays() static method.
+    They should be accessed using the K8090Traits::Command enum values.
+    Example, which shows how to build the whole command _query relay status_:
     \code
-    QString strCmd(commandBase);
+    QString strCmd(kStxByte);
     int n = 0; // number of data bytes
-    strCmd.append(strCommands[static_cast<int>(Command::ReadStatus)])
-          .append(QString(" %1 ").arg(n, 2, 16, QChar('0')).toUpper());
-    strCmd.append(checkSum(strCmd));
+    strCmd.append(strCommands[static_cast<int>(Command::RELAY_STATUS)])
+          .append(" 00 00 00 ")
+          .append(checkSum(strCmd))
+          .append(QString(" %1".arg(kEtxByte_)));
     \endcode
 */
-QString K8090::strCommands[static_cast<int>(Command::None)];
+QString K8090::str_commands_[static_cast<int>(Command::NONE)];
+
 
 /*!
   \brief Creates a new K8090 instance and sets the default values.
@@ -124,12 +127,13 @@ QString K8090::strCommands[static_cast<int>(Command::None)];
 K8090::K8090(QObject *parent) :
     QObject(parent)
 {
-    lastCommand = Command::None;
+    last_command_ = Command::NONE;
     fillCommandsArrays();
 
-    serialPort_ = new QSerialPort(this);
-    connect(serialPort_, &QSerialPort::readyRead, this, &K8090::onReadyData);
+    serial_port_ = new QSerialPort(this);
+    connect(serial_port_, &QSerialPort::readyRead, this, &K8090::onReadyData);
 }
+
 
 QList<ComPortParams> K8090::availablePorts()
 {
@@ -146,15 +150,16 @@ QList<ComPortParams> K8090::availablePorts()
     return comPortParamsList;
 }
 
+
 void K8090::connectK8090()
 {
     bool cardFound = false;
     foreach (const QSerialPortInfo &info, QSerialPortInfo::availablePorts()) {  // NOLINT(whitespace/parens)
-        if (info.productIdentifier() == productID &&
-                info.vendorIdentifier() == vendorID) {
+        if (info.productIdentifier() == kProductID &&
+                info.vendorIdentifier() == kVendorID) {
             cardFound = true;
-            comPortName_ = info.portName();
-            qDebug() << "Port name: " % comPortName_;
+            com_port_name_ = info.portName();
+            qDebug() << "Port name: " % com_port_name_;
         }
     }
 
@@ -163,15 +168,16 @@ void K8090::connectK8090()
         return;
     }
 
-    serialPort_->setPortName(comPortName_);
-    serialPort_->setBaudRate(QSerialPort::Baud19200);
-    serialPort_->setDataBits(QSerialPort::Data8);
-    serialPort_->setParity(QSerialPort::NoParity);
-    serialPort_->setStopBits(QSerialPort::OneStop);
-    serialPort_->setFlowControl(QSerialPort::NoFlowControl);
+    serial_port_->setPortName(com_port_name_);
+    serial_port_->setBaudRate(QSerialPort::Baud19200);
+    serial_port_->setDataBits(QSerialPort::Data8);
+    serial_port_->setParity(QSerialPort::NoParity);
+    serial_port_->setStopBits(QSerialPort::OneStop);
+    serial_port_->setFlowControl(QSerialPort::NoFlowControl);
 
     sendSwitchRelayOffCommand();
 }
+
 
 /*!
    \brief K8090::onSendToSerial
@@ -181,24 +187,26 @@ void K8090::connectK8090()
 void K8090::onSendToSerial(const unsigned char *buffer, int n)
 {
     qDebug() << byteToHex(buffer, n);
-    if (!serialPort_->isOpen())
-        serialPort_->open(QIODevice::ReadWrite);
-    serialPort_->write(reinterpret_cast<char*>(const_cast<unsigned char*>(buffer)), n);
+    if (!serial_port_->isOpen())
+        serial_port_->open(QIODevice::ReadWrite);
+    serial_port_->write(reinterpret_cast<char*>(const_cast<unsigned char*>(buffer)), n);
     delete[] buffer;
 }
+
 
 void K8090::onReadyData()
 {
     qDebug() << "R8090::onReadyData";
 
     // converting the data to unsigned char
-    QByteArray data = serialPort_->readAll();
+    QByteArray data = serial_port_->readAll();
     int n = data.size();
     unsigned char *buffer = reinterpret_cast<unsigned char*>(data.data());
     qDebug() << byteToHex(buffer, n);
 
-    lastCommand = Command::None;
+    last_command_ = Command::NONE;
 }
+
 
 /*!
     \brief K8090::sendCommand
@@ -206,6 +214,7 @@ void K8090::onReadyData()
 void K8090::sendCommand()
 {
 }
+
 
 /*!
    \brief K8090::sendSwitchRelayOnCommand
@@ -216,14 +225,15 @@ void K8090::sendSwitchRelayOnCommand()
     unsigned char * cmd = new unsigned char[n];
     int ii;
     for (ii = 0; ii < 2; ++ii)
-        cmd[ii] = bCommands[static_cast<int>(Command::SwitchRelayOn)][ii];
-    cmd[2] = (unsigned char) RelayID::One;
+        cmd[ii] = b_commands_[static_cast<int>(Command::RELAY_ON)][ii];
+    cmd[2] = (unsigned char) RelayID::ONE;
     cmd[5] = checkSum(cmd, 5);
-    cmd[6] = bEtxByte;
-    lastCommand = Command::SwitchRelayOn;
+    cmd[6] = b_etx_byte_;
+    last_command_ = Command::RELAY_ON;
     qDebug() << byteToHex(cmd, n);
     onSendToSerial(cmd, n);
 }
+
 
 /*!
    \brief K8090::sendSwitchRelayOnCommand
@@ -234,14 +244,15 @@ void K8090::sendSwitchRelayOffCommand()
     unsigned char * cmd = new unsigned char[n];
     int ii;
     for (ii = 0; ii < 2; ++ii)
-        cmd[ii] = bCommands[static_cast<int>(Command::SwitchRelayOff)][ii];
-    cmd[2] = (unsigned char) RelayID::One;
+        cmd[ii] = b_commands_[static_cast<int>(Command::RELAY_OFF)][ii];
+    cmd[2] = (unsigned char) RelayID::ONE;
     cmd[5] = checkSum(cmd, 5);
-    cmd[6] = bEtxByte;
-    lastCommand = Command::SwitchRelayOff;
+    cmd[6] = b_etx_byte_;
+    last_command_ = Command::RELAY_OFF;
     qDebug() << byteToHex(cmd, n);
     onSendToSerial(cmd, n);
 }
+
 
 /*!
    \brief K8090::hexToByte
@@ -271,6 +282,7 @@ void K8090::hexToByte(unsigned char **pbuffer, int *n, const QString &msg)
     }
 }
 
+
 /*!
    \brief K8090::byteToHex
    \param buffer
@@ -288,6 +300,7 @@ QString K8090::byteToHex(const unsigned char *buffer, int n)
     }
     return msg;
 }
+
 
 /*!
    \brief K8090::checkSum
@@ -307,6 +320,7 @@ QString K8090::checkSum(const QString &msg)
     return QString("%1").arg((unsigned int)bChk, 2, 16, QChar('0')).toUpper();
 }
 
+
 /*!
    \brief K8090::checkSum
    \param bMsg
@@ -325,6 +339,7 @@ unsigned char K8090::checkSum(const unsigned char *bMsg, int n)
 
     return byteSum;
 }
+
 
 /*!
    \brief K8090::validateResponse
@@ -346,6 +361,7 @@ bool K8090::validateResponse(const QString &msg, Command cmd)
     }
 }
 
+
 /*!
     \brief K8090::validateResponse
     \param bMsg Pointer to field of bytes containing the response.
@@ -357,7 +373,7 @@ bool K8090::validateResponse(const unsigned char *bMsg, int n, Command cmd)
 {
     if (n >= 6) {
         for (int ii = 0; ii < 2; ++ii)
-            if (bMsg[ii] != bCommands[static_cast<int>(cmd)][ii])
+            if (bMsg[ii] != b_commands_[static_cast<int>(cmd)][ii])
                 return false;
         unsigned char bTest[5];
         for (int ii = 0; ii < 5; ++ii)
@@ -369,40 +385,42 @@ bool K8090::validateResponse(const unsigned char *bMsg, int n, Command cmd)
     return false;
 }
 
+
 /*!
    \brief Fills arrays containing commands according to their importance.
  */
 void K8090::fillCommandsArrays()
 {
-    strCommands[static_cast<int>(Command::SwitchRelayOn)] = switchRelayOnCmd;
-    strCommands[static_cast<int>(Command::SwitchRelayOff)] = switchRelayOffCmd;
-    strCommands[static_cast<int>(Command::ToggleRelay)] = toggleRelayCmd;
-    strCommands[static_cast<int>(Command::QueryRelayStatus)] = queryRelayStatusCmd;
-    strCommands[static_cast<int>(Command::SetButtonMode)] = setButtonModeCmd;
-    strCommands[static_cast<int>(Command::QueryButtonMode)] = queryButtonModeCmd;
-    strCommands[static_cast<int>(Command::StartRelayTimer)] = startRelayTimerCmd;
-    strCommands[static_cast<int>(Command::SetRelayTimerDelay)] = setRelayTimerDelayCmd;
-    strCommands[static_cast<int>(Command::QueryTimerDelay)] = queryTimerDelayCmd;
-    strCommands[static_cast<int>(Command::ButtonStatus)] = buttonStatusCmd;
-    strCommands[static_cast<int>(Command::RelayStatus)] = relayStatusCmd;
-    strCommands[static_cast<int>(Command::ResetFactoryDefaults)] = resetFactoryDefaultsCmd;
-    strCommands[static_cast<int>(Command::JumperStatus)] = jumperStatusCmd;
-    strCommands[static_cast<int>(Command::FirmwareVersion)] = firmwareVersionCmd;
+    str_commands_[static_cast<int>(Command::RELAY_ON)] = kSwitchRelayOnCmd_;
+    str_commands_[static_cast<int>(Command::RELAY_OFF)] = kSwitchRelayOffCmd_;
+    str_commands_[static_cast<int>(Command::TOGGLE_RELAY)] = kToggleRelayCmd_;
+    str_commands_[static_cast<int>(Command::RELAY_STATUS)] = kQueryRelayStatusCmd_;
+    str_commands_[static_cast<int>(Command::SET_BUTTON_MODE)] = kSetButtonModeCmd_;
+    str_commands_[static_cast<int>(Command::BUTTON_MODE)] = kQueryButtonModeCmd_;
+    str_commands_[static_cast<int>(Command::START_TIMER)] = kStartRelayTimerCmd_;
+    str_commands_[static_cast<int>(Command::SET_TIMER)] = kSetRelayTimerDelayCmd_;
+    str_commands_[static_cast<int>(Command::TIMER)] = kQueryTimerDelayCmd_;
+    str_commands_[static_cast<int>(Command::BUTTON_STATUS)] = kButtonStatusCmd_;
+    str_commands_[static_cast<int>(Command::RELAY_STATUS)] = kRelayStatusCmd_;
+    str_commands_[static_cast<int>(Command::RESET_FACTORY_DEFAULTS)] = kResetFactoryDefaultsCmd_;
+    str_commands_[static_cast<int>(Command::JUMPER_STATUS)] = kJumperStatusCmd_;
+    str_commands_[static_cast<int>(Command::FIRMWARE_VERSION)] = kFirmwareVersionCmd_;
 
     bool ok;
-    for (int ii = 0; ii < static_cast<int>(Command::None); ++ii) {
-        bCommands[ii][0] = stxByte.toUInt(&ok, 16);
-        bCommands[ii][1] = static_cast<unsigned char>(strCommands[ii].toUInt(&ok, 16));
+    for (int ii = 0; ii < static_cast<int>(Command::NONE); ++ii) {
+        b_commands_[ii][0] = kStxByte_.toUInt(&ok, 16);
+        b_commands_[ii][1] = static_cast<unsigned char>(str_commands_[ii].toUInt(&ok, 16));
     }
 
-    bEtxByte = etxByte.toUInt(&ok, 16);
+    b_etx_byte_ = kEtxByte_.toUInt(&ok, 16);
 }
+
 
 /*!
    \brief K8090::~K8090
  */
 K8090::~K8090()
 {
-    serialPort_->close();
-    delete serialPort_;
+    serial_port_->close();
+    delete serial_port_;
 }