+++ src/sprelay/core/k8090.h
@@ -30,40 +30,43 @@
 // forward declarations
 class QSerialPort;
 
+
 namespace K8090Traits
 {
 enum class Command {
-    SwitchRelayOn,
-    SwitchRelayOff,
-    ToggleRelay,
-    QueryRelayStatus,
-    SetButtonMode,
-    QueryButtonMode,
-    StartRelayTimer,
-    SetRelayTimerDelay,
-    QueryTimerDelay,
-    ButtonStatus,
-    RelayStatus,
-    ResetFactoryDefaults,
-    JumperStatus,
-    FirmwareVersion,
-    None
+    RELAY_ON,
+    RELAY_OFF,
+    TOGGLE_RELAY,
+    QUERY_RELAY,
+    SET_BUTTON_MODE,
+    BUTTON_MODE,
+    START_TIMER,
+    SET_TIMER,
+    TIMER,
+    BUTTON_STATUS,
+    RELAY_STATUS,
+    RESET_FACTORY_DEFAULTS,
+    JUMPER_STATUS,
+    FIRMWARE_VERSION,
+    NONE
 };
 
 inline Command& operator++(Command &cmd) { cmd = static_cast<Command>(static_cast<int>(cmd) + 1); return cmd; }
 inline Command operator++(Command &cmd, int) { Command tmp(cmd); cmd++; return tmp; }
 
+
 enum struct RelayID : unsigned char
 {
-    None  = 0,
-    One   = 1 << 0,
-    Two   = 1 << 1,
-    Three = 1 << 2,
-    Four  = 1 << 3,
-    Five  = 1 << 4,
-    Six   = 1 << 5,
-    Seven = 1 << 6,
-    Eight = 1 << 7
+    NONE  = 0,
+    ONE   = 1 << 0,
+    TWO   = 1 << 1,
+    THREE = 1 << 2,
+    FOUR  = 1 << 3,
+    FIVE  = 1 << 4,
+    SIX   = 1 << 5,
+    SEVEN = 1 << 6,
+    EIGHT = 1 << 7,
+    ALL   = 0xff
 };
 
 
@@ -85,7 +88,11 @@ struct ComPortParams
 class K8090 : public QObject
 {
     Q_OBJECT
+
 public:  // NOLINT(whitespace/indent)
+    static const quint16 kProductID;
+    static const quint16 kVendorID;
+
     explicit K8090(QObject *parent = 0);
     virtual ~K8090();
 
@@ -94,60 +101,60 @@ public:  // NOLINT(whitespace/indent)
 
 signals:  // NOLINT(whitespace/indent)
 
+
 public slots:  // NOLINT(whitespace/indent)
     void connectK8090();
+
 
 protected slots:  // NOLINT(whitespace/indent)
     void onSendToSerial(const unsigned char *buffer, int n);
 
+
 private slots:  // NOLINT(whitespace/indent)
     void onReadyData();
 
+
 private:  // NOLINT(whitespace/indent)
-    void sendCommand();
-
-    void sendSwitchRelayOnCommand();
-    void sendSwitchRelayOffCommand();
-
-    static const quint16 productID;
-    static const quint16 vendorID;
-
-    QString comPortName_;
-    QSerialPort *serialPort_;
-
     static void hexToByte(unsigned char **pbuffer, int *n, const QString &msg);
     static QString byteToHex(const unsigned char *buffer, int n);
     static QString checkSum(const QString &msg);
     static unsigned char checkSum(const unsigned char *bMsg, int n);
     static bool validateResponse(const QString &msg, K8090Traits::Command cmd);
     static bool validateResponse(const unsigned char *bMsg, int n, K8090Traits::Command cmd);
-
     static void fillCommandsArrays();
 
-    bool commandBuffer[static_cast<int>(K8090Traits::Command::None)];
+    void sendCommand();
 
-    K8090Traits::Command lastCommand;
+    void sendSwitchRelayOnCommand();
+    void sendSwitchRelayOffCommand();
 
-    static unsigned char bCommands[static_cast<int>(K8090Traits::Command::None)][2];
-    static unsigned char bEtxByte;
-    static QString strCommands[static_cast<int>(K8090Traits::Command::None)];
+    QString com_port_name_;
+    QSerialPort *serial_port_;
 
-    static const QString stxByte;
-    static const QString etxByte;
-    static const QString switchRelayOnCmd;
-    static const QString switchRelayOffCmd;
-    static const QString toggleRelayCmd;
-    static const QString queryRelayStatusCmd;
-    static const QString setButtonModeCmd;
-    static const QString queryButtonModeCmd;
-    static const QString startRelayTimerCmd;
-    static const QString setRelayTimerDelayCmd;
-    static const QString queryTimerDelayCmd;
-    static const QString buttonStatusCmd;
-    static const QString relayStatusCmd;
-    static const QString resetFactoryDefaultsCmd;
-    static const QString jumperStatusCmd;
-    static const QString firmwareVersionCmd;
+    bool commandBuffer_[static_cast<int>(K8090Traits::Command::NONE)];
+
+    K8090Traits::Command last_command_;
+
+    static unsigned char b_commands_[static_cast<int>(K8090Traits::Command::NONE)][2];
+    static unsigned char b_etx_byte_;
+    static QString str_commands_[static_cast<int>(K8090Traits::Command::NONE)];
+
+    static const QString kStxByte_;
+    static const QString kEtxByte_;
+    static const QString kSwitchRelayOnCmd_;
+    static const QString kSwitchRelayOffCmd_;
+    static const QString kToggleRelayCmd_;
+    static const QString kQueryRelayStatusCmd_;
+    static const QString kSetButtonModeCmd_;
+    static const QString kQueryButtonModeCmd_;
+    static const QString kStartRelayTimerCmd_;
+    static const QString kSetRelayTimerDelayCmd_;
+    static const QString kQueryTimerDelayCmd_;
+    static const QString kButtonStatusCmd_;
+    static const QString kRelayStatusCmd_;
+    static const QString kResetFactoryDefaultsCmd_;
+    static const QString kJumperStatusCmd_;
+    static const QString kFirmwareVersionCmd_;
 };
 
 #endif  // SPRELAY_CORE_K8090_H_