+++ src/sprelay/gui/mainwindow.cpp
@@ -22,84 +22,82 @@
 
 #include "mainwindow.h"
 
-#include <QPushButton>
-#include <QLabel>
 #include <QComboBox>
+#include <QDebug>
+#include <QLabel>
 #include <QLayout>
 #include <QMessageBox>
+#include <QPushButton>
 #include <QStringBuilder>
 
-#include <QDebug>
-
 #include "k8090.h"
+
 
 MainWindow::MainWindow()
 {
     connected_ = false;
 
-    cWidget = new QWidget;
+    central_widget_ = new QWidget(this);
 
-    connectButton = new QPushButton(tr("Connect"));
-    setCentralWidget(cWidget);
+    connect_button_ = new QPushButton(tr("Connect"), this);
+    setCentralWidget(central_widget_);
 
-    qDebug() << "Ahoj";
+    k8090_ = new K8090(this);
 
-    k8090 = new K8090(this);
-
-    refreshPortsButton = new QPushButton(tr("Refresh Ports"));
-    portsLabel = new QLabel(tr("Select port:"));
-    portsComboBox = new QComboBox();
-    portsLabel->setBuddy(portsComboBox);  // buddy accepts focus instead of label (for editing)
+    refresh_ports_button_ = new QPushButton(tr("Refresh Ports"), this);
+    ports_label_ = new QLabel(tr("Select port:"), this);
+    ports_combo_box_ = new QComboBox(this);
+    ports_label_->setBuddy(ports_combo_box_);  // buddy accepts focus instead of label (for editing)
     int index = 0;
 //    refreshingPortsComboBoxContent = true;
     foreach (const K8090Traits::ComPortParams &comPortParams, K8090::availablePorts()) {  // NOLINT(whitespace/parens)
-        portsComboBox->insertItem(index++, comPortParams.portName);
+        ports_combo_box_->insertItem(index++, comPortParams.portName);
     }
     if (connected_) {
 //        comPortName_ = k8090->comPortName();
     }
-    if ((index = portsComboBox->findText(comPortName_)) != -1)
-        portsComboBox->setCurrentIndex(index);
+    if ((index = ports_combo_box_->findText(com_port_name_)) != -1)
+        ports_combo_box_->setCurrentIndex(index);
 //    refreshingPortsComboBoxContent = false;
     if (!connected_ || index == -1) {
 //        k8090->setComPortName(portsComboBox->currentText());
     }
-    comPortName_ = portsComboBox->currentText();
+    com_port_name_ = ports_combo_box_->currentText();
 
-    connect(connectButton, &QPushButton::clicked, this, &MainWindow::onConnectButtonClicked);
-    connect(refreshPortsButton, &QPushButton::clicked, this, &MainWindow::onRefreshPortsButtonClicked);
-    connect(portsComboBox, static_cast<void (QComboBox::*)(const QString &)>(&QComboBox::currentIndexChanged),
+    connect(connect_button_, &QPushButton::clicked, this, &MainWindow::onConnectButtonClicked);
+    connect(refresh_ports_button_, &QPushButton::clicked, this, &MainWindow::onRefreshPortsButtonClicked);
+    connect(ports_combo_box_, static_cast<void (QComboBox::*)(const QString &)>(&QComboBox::currentIndexChanged),
             this, &MainWindow::onPortsComboBoxCurrentIndexChanged);
 
-    mainHLayout = new QHBoxLayout;
-    cWidget->setLayout(mainHLayout);
+    main_layout_ = new QHBoxLayout;
+    central_widget_->setLayout(main_layout_);
 
-    vLayout1 = new QVBoxLayout;
-    mainHLayout->addLayout(vLayout1);
-    vLayout1->addWidget(connectButton);
-    vLayout1->addStretch();
+    v_layout_1 = new QVBoxLayout;
+    main_layout_->addLayout(v_layout_1);
+    v_layout_1->addWidget(connect_button_);
+    v_layout_1->addStretch();
 
-    vLayout2 = new QVBoxLayout;
-    mainHLayout->addLayout(vLayout2);
+    v_layout_2 = new QVBoxLayout;
+    main_layout_->addLayout(v_layout_2);
 
-    refreshPortsHLayout = new QHBoxLayout;
-    vLayout2->addLayout(refreshPortsHLayout);
-    refreshPortsHLayout->addStretch();
-    refreshPortsHLayout->addWidget(refreshPortsButton);
-    refreshPortsHLayout->addStretch();
+    refresh_ports_h_layout_ = new QHBoxLayout;
+    v_layout_2->addLayout(refresh_ports_h_layout_);
+    refresh_ports_h_layout_->addStretch();
+    refresh_ports_h_layout_->addWidget(refresh_ports_button_);
+    refresh_ports_h_layout_->addStretch();
 
-    portsHLayout = new QHBoxLayout;
-    vLayout2->addLayout(portsHLayout);
-    portsHLayout->addWidget(portsLabel);
-    portsHLayout->addWidget(portsComboBox);
-    portsHLayout->addStretch();
+    ports_h_layout_ = new QHBoxLayout;
+    v_layout_2->addLayout(ports_h_layout_);
+    ports_h_layout_->addWidget(ports_label_);
+    ports_h_layout_->addWidget(ports_combo_box_);
+    ports_h_layout_->addStretch();
 
-    vLayout2->addStretch();
+    v_layout_2->addStretch();
 }
 
 void MainWindow::onConnectButtonClicked()
 {
-    k8090->connectK8090();
+    k8090_->connectK8090();
 }
 
 
@@ -110,9 +108,9 @@ void MainWindow::onPortsComboBoxCurrentIndexChanged(const QString &portName)
 
 void MainWindow::onRefreshPortsButtonClicked()
 {
-    if (!k8090->availablePorts().isEmpty()) {
+    if (!k8090_->availablePorts().isEmpty()) {
         QString msg;
-        QString currPort = portsComboBox->currentText();
+        QString currPort = ports_combo_box_->currentText();
         QStringList comPortNames;
         foreach (const K8090Traits::ComPortParams &comPortParams, K8090::availablePorts()) {  // NOLINT
             msg.append("Port name: " % comPortParams.portName % "\n" %
@@ -125,27 +123,27 @@ void MainWindow::onRefreshPortsButtonClicked()
         QMessageBox::information(this, tr("Serial ports information:"), msg, QMessageBox::Ok);
         bool ok = true;
         bool counteq = false;
-        if ((counteq = (portsComboBox->count() == comPortNames.count()))) {
+        if ((counteq = (ports_combo_box_->count() == comPortNames.count()))) {
             for (int ii = 0; ii < comPortNames.count(); ++ii) {
-                if (portsComboBox->findText(comPortNames.at(ii)) < 0) {
+                if (ports_combo_box_->findText(comPortNames.at(ii)) < 0) {
                     ok = false;
                 }
             }
         }
         if (!counteq || !ok) {
 //            refreshingPortsComboBoxContent = true;
-            portsComboBox->clear();
-            portsComboBox->insertItems(1, comPortNames);
+            ports_combo_box_->clear();
+            ports_combo_box_->insertItems(1, comPortNames);
             int index;
-            if ((index = portsComboBox->findText(currPort)) >= 0)
-                portsComboBox->setCurrentIndex(index);
-            else if ((index = portsComboBox->findText(comPortName_)) >= 0)
-                portsComboBox->setCurrentIndex(index);
+            if ((index = ports_combo_box_->findText(currPort)) >= 0)
+                ports_combo_box_->setCurrentIndex(index);
+            else if ((index = ports_combo_box_->findText(com_port_name_)) >= 0)
+                ports_combo_box_->setCurrentIndex(index);
 //            refreshingPortsComboBoxContent = false;
             if (!connected_ && index == -1) {
 //                neslab->setComPortName(portsComboBox->currentText());
             }
-            comPortName_ = portsComboBox->currentText();
+            com_port_name_ = ports_combo_box_->currentText();
         }
     } else {
 //        onDisconnected();